# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ownEvo Kernel API
  version: 0.1.0
  description: |
    REST + Server-Sent Events contract between the Python kernel (`apps/kernel/`)
    and the TS web app (`apps/web/`). Locked 2026-05-03 by eng review (D2).

    Reference artifacts (hand-maintained, not auto-generated):
      - Python Pydantic models: `apps/kernel/src/ownevo_kernel/api/models.py`
      - TypeScript types: `apps/web/lib/api.ts` (hand-maintained; no codegen step)

    Auth model (MVP, single-tenant per D4): no auth in development;
    dev-token middleware in staging. Production auth is Phase 2.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/workflows:
    post:
      tags: [workflows]
      summary: Create workflow from natural-language description (NL-gen)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Created. Server begins NL-gen + meta-eval; subscribe to /api/workflows/{id}/stream for progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
    get:
      tags: [workflows]
      summary: List workflows with summary metrics (W7 slice 2 — Health page)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
...
  /api/workflows/{id}:
    patch:
      tags: [workflows]
      summary: Edit the workflow description
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAnatomyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [workflows]
      summary: Hard-delete a workflow and every domain row tied to it
      description: |
        Cascades in FK-safe order inside one transaction. Audit entries are
        intentionally NOT deleted (append-only WORM per D2).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deletion summary with row counts per table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/workflows/{workflow_id}/skills:
    get:
      tags: [skills]
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
  /api/skills/{skill_id}:
    get:
      tags: [skills]
      parameters:
        - name: skill_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetailResponse'

  # ---------- clusters ----------
...
components:
  schemas:
...
    WorkflowMode:
      type: string
      enum: [eval-only, eval-propose, gated, autonomous]
...
    Workflow:
      type: object
      required: [id, description, spec, created_at]
      properties:
        id: { type: string }
        description: { type: string }
        spec:
          type: object
          description: Frozen-schema NL-gen output (W3 freeze). See packages/trace-format/ for shape.
        metric_id: { type: string }
        sim_skill_id: { type: string }
        meta_eval_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        mode: { $ref: '#/components/schemas/WorkflowMode' }
        created_at: { type: string, format: date-time }
    CreateWorkflowRequest:
      type: object
      required: [description]
      properties:
        description: { type: string, minLength: 10 }

    # W7 slice 2 — Health page row + LiftChart point
    WorkflowSummary:
      type: object
      required:
        - id
        - description
        - mode
        - iteration_count
        - pending_proposals_count
      properties:
        id: { type: string }
        description: { type: string }
        mode: { $ref: '#/components/schemas/WorkflowMode' }
        kind: { type: string, nullable: true, description: "Workflow kind tag (e.g. 'nl-gen'). Null on legacy rows." }
        iteration_count: { type: integer }
        running_iteration_count: { type: integer, description: "Number of iterations currently in 'running' state." }
        oldest_running_started_at: { type: string, format: date-time, nullable: true, description: "started_at of the oldest running iteration; null when none." }
        best_ever_score: { type: number, format: float, nullable: true }
        last_improved_at: { type: string, format: date-time, nullable: true }
        pending_proposals_count: { type: integer }
    WorkflowListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/WorkflowSummary' }
        total: { type: integer }
...
    WorkflowAnatomyResponse:
      type: object
      required: [id, description, mode, spec]
      properties:
        id: { type: string }
        description: { type: string }
        mode: { $ref: '#/components/schemas/WorkflowMode' }
        kind: { type: string, nullable: true }
        spec:
          type: object
          description: |
            Raw NL-gen spec JSONB. Conforms to
            packages/kernel/src/ownevo_kernel/nl_gen/spec.py:WorkflowSpec
            v1.1 — keys: domain, environment, tools, reviewer,
            success_criterion, ui (ui.tabs[].primitives now includes
            ScheduleGrid in addition to the original 8 types).

    # W7 slices 9 + 10 (7.1.10 + 7.1.11) — skill detail DTOs
    SkillSummaryResponse:
      type: object
      required: [id, kind, capability_tags]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [python, instruction, composite]
        workflow_id: { type: string, nullable: true }
        capability_tags:
          type: array
          items: { type: string }
        head_version_id: { type: string, format: uuid, nullable: true }
        head_version_seq: { type: integer, nullable: true }
        head_created_at: { type: string, format: date-time, nullable: true }
    SkillListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SkillSummaryResponse' }
    SkillVersionSummary:
      type: object
      required: [id, version_seq, created_by, created_at]
      properties:
        id: { type: string, format: uuid }
        version_seq: { type: integer }
        parent_version_id: { type: string, format: uuid, nullable: true }
        diff_summary: { type: string, nullable: true }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    SkillRelatedEvalCase:
      type: object
      required: [id, provenance, is_test_fold, created_at]
      properties:
        id: { type: string, format: uuid }
        workflow_id: { type: string, nullable: true }
        provenance: { type: string }
        expected_behavior: { type: object, nullable: true }
        is_test_fold: { type: boolean }
        created_at: { type: string, format: date-time }
    SkillDetailResponse:
      type: object
      required: [id, kind, capability_tags, versions, related_eval_cases]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [python, instruction, composite]
        workflow_id: { type: string, nullable: true }
        workflow_description: { type: string, nullable: true }
        capability_tags:
          type: array
          items: { type: string }
        head_version_id: { type: string, format: uuid, nullable: true }
        head_version_seq: { type: integer, nullable: true }
        head_content: { type: string, nullable: true }
        head_retention_block: { type: object, nullable: true }
        head_diff_summary: { type: string, nullable: true }
        head_created_at: { type: string, format: date-time, nullable: true }
        head_created_by: { type: string, nullable: true }
        parent_content: { type: string, nullable: true }
        parent_version_seq: { type: integer, nullable: true }
        versions:
          type: array
          items: { $ref: '#/components/schemas/SkillVersionSummary' }
        related_eval_cases:
          type: array
          items: { $ref: '#/components/schemas/SkillRelatedEvalCase' }
...
    WorkflowUpdateRequest:
      type: object
      required: [description]
      properties:
        description: { type: string, minLength: 1 }
    WorkflowDeleteResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
        approvals: { type: integer }
        learnings: { type: integer }
        proposals: { type: integer }
        traces: { type: integer }
        meta_evals: { type: integer }
        iterations: { type: integer }
        eval_cases: { type: integer }
        failure_clusters: { type: integer }
        skill_versions: { type: integer }
        skills: { type: integer }

    # ---------- SSE event types ----------
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'